+++ QuantConnect.AlphaVantage.Tests/config.json
@@ -2,5 +2,5 @@
   "data-folder": "../../../../Lean/Data/",
   "data-directory": "../../../../Lean/Data/",
 
-  "alpha-vantage-api-key":  ""
+  "alpha-vantage-api-key": ""
 }
